+++ Source/MQTTnet/Server/MqttRetainedMessageMatch.cs
@@ -2,13 +2,20 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using MQTTnet.Protocol;
 
 namespace MQTTnet.Server
 {
     public sealed class MqttRetainedMessageMatch
     {
-        public MqttApplicationMessage ApplicationMessage { get; set; }
+        public MqttRetainedMessageMatch(MqttApplicationMessage applicationMessage, MqttQualityOfServiceLevel subscriptionQualityOfServiceLevel)
+        {
+            ApplicationMessage = applicationMessage ?? throw new ArgumentNullException(nameof(applicationMessage));
+            SubscriptionQualityOfServiceLevel = subscriptionQualityOfServiceLevel;
+        }
+
+        public MqttApplicationMessage ApplicationMessage { get; }
 
         public MqttQualityOfServiceLevel SubscriptionQualityOfServiceLevel { get; set; }
     }